- IpfixDbWriter: bugfix - IpfixReceiverSctp: select with timeout, code cleanup - IpfixReceiverUdp: code cleanup - PSAMPExporterCfg, IpfixCollectorCfg: enable SCTP export, enable observerationDomainId configuration - IpfixParser: convert to unix file format

+++ branches/vermont/anonym/concentrator/IpfixSender.cpp
@@ -46,7 +46,7 @@ using namespace std;
  * @param port destination collector's port
  * @return handle to use when calling @c destroyIpfixSender()
  */
-IpfixSender::IpfixSender(uint16_t observationDomainId, uint32_t maxRecordRate, uint32_t sctpDataLifetime, uint32_t sctpReconnectInterval,
+IpfixSender::IpfixSender(uint32_t observationDomainId, uint32_t maxRecordRate, uint32_t sctpDataLifetime, uint32_t sctpReconnectInterval,
 		uint32_t templateRefreshInterval, uint32_t templateRefreshRate)
 	: statSentPackets(0),
 	  noCachedRecords(0),
